Rename frizzle to connectors
