+++ f/connectors/comapeo/comapeo_observations.script.yaml
@@ -1,6 +1,6 @@
 summary: 'CoMapeo: Fetch Observations'
 description: This script fetches observations and attachments from the REST API of a CoMapeo archive server.
-lock: '!inline f/frizzle/comapeo/comapeo_observations.script.lock'
+lock: '!inline f/connectors/comapeo/comapeo_observations.script.lock'
 concurrency_time_window_s: 0
 kind: script
 schema:
