+++ f/connectors/comapeo/tests/assets/capybara.jpg
