+++ f/connectors/folder.meta.yaml
@@ -1,0 +1,2 @@
+summary: null
+display_name: connectors
