+++ f/connectors/kobotoolbox/tests/assets/trees.png
