+++ f/connectors/kobotoolbox/tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 import responses
 import testing.postgresql
 
-from f.frizzle.kobotoolbox.tests.assets import server_responses
+from f.connectors.kobotoolbox.tests.assets import server_responses
 
 
 @pytest.fixture
@@ -40,7 +40,7 @@ def koboserver(mocked_responses):
     )
     mocked_responses.get(
         re.compile(rf"{server_url}/api/v2/assets/{form_id}/data/\d+/attachments/\d+/?"),
-        body=open("f/frizzle/kobotoolbox/tests/assets/trees.png", "rb").read(),
+        body=open("f/connectors/kobotoolbox/tests/assets/trees.png", "rb").read(),
         content_type="image/png",
         headers={"Content-Length": "3632"},
     )
