+++ f/connectors/kobotoolbox/tests/kobotoolboxresponses_test.py
@@ -1,6 +1,6 @@
 import psycopg2
 
-from f.frizzle.kobotoolbox.kobotoolbox_responses import main, sanitize
+from f.connectors.kobotoolbox.kobotoolbox_responses import main, sanitize
 
 
 def test_sanitize():
