--- f/frizzle/README.md
@@ -1,7 +1,0 @@
-# Frizzle: The Magic Data Bus
-
-Frizzle is a collection of lightweight and flexible scripts, designed to enable seamless integration of multiple data sources.
-
-These Frizzle scripts can fetch data from various upstream sources, such as KoboToolbox, CoMapeo, or Google Cloud Platform. Data from these sources is ingested in a consistent way. The scripts ensure that the data is cleansed, validated, and formatted appropriately before being transferred to its target destination, such as a SQL database, file storage, or other handling methods.
-
-Using Windmill, the scripts can be scheduled to run at defined intervals, ensuring timely data processing.
