--- f/frizzle/folder.meta.yaml
@@ -1,6 +1,0 @@
-summary: null
-display_name: frizzle
-extra_perms:
-  u/jeffgerard: true
-owners:
-  - u/jeffgerard
