+++ tox.ini
@@ -8,12 +8,12 @@ setenv =
 
 [testenv:alerts]
 deps =
-    -r{toxinidir}/f/frizzle/alerts/alerts_gcs.script.lock
-    -r{toxinidir}/f/frizzle/alerts/tests/requirements-test.txt
+    -r{toxinidir}/f/connectors/alerts/alerts_gcs.script.lock
+    -r{toxinidir}/f/connectors/alerts/tests/requirements-test.txt
 docker =
     gcp-storage-emulator
 commands =
-    pytest {posargs} f/frizzle/alerts
+    pytest {posargs} f/connectors/alerts
 
 [docker:gcp-storage-emulator]
 image = oittaa/gcp-storage-emulator
@@ -24,17 +24,17 @@ expose =
 
 [testenv:comapeo_observations]
 deps =
-    -r{toxinidir}/f/frizzle/comapeo/comapeo_observations.script.lock
-    -r{toxinidir}/f/frizzle/comapeo/tests/requirements-test.txt
+    -r{toxinidir}/f/connectors/comapeo/comapeo_observations.script.lock
+    -r{toxinidir}/f/connectors/comapeo/tests/requirements-test.txt
 commands =
-    pytest {posargs} f/frizzle/comapeo
+    pytest {posargs} f/connectors/comapeo
 
 [testenv:kobotoolbox]
 deps =
-    -r{toxinidir}/f/frizzle/kobotoolbox/kobotoolbox_responses.script.lock
-    -r{toxinidir}/f/frizzle/kobotoolbox/tests/requirements-test.txt
+    -r{toxinidir}/f/connectors/kobotoolbox/kobotoolbox_responses.script.lock
+    -r{toxinidir}/f/connectors/kobotoolbox/tests/requirements-test.txt
 commands =
-    pytest {posargs} f/frizzle/kobotoolbox
+    pytest {posargs} f/connectors/kobotoolbox
 
 [testenv:postgres_to_geojson]
 deps =
